% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{Analyser_fromDataframe}
\alias{Analyser_fromDataframe}
\title{Create an analyser from a data frame}
\usage{
Analyser$fromSearcher(searcher)
}
\arguments{
\item{omop}{an instance of the Omop class}

\item{groupedDf}{a data frame grouped to define the document level groupings in the data}

\item{field}{the field that holds the concept_id}

\item{expanded}{has the searcher already had ancestors expanded?}
}
\value{
an analyser with idf and graph counts precomputed.
}
\description{
creates a network analyser from a searcher
}
