% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{Analyser_fromSearcher}
\alias{Analyser_fromSearcher}
\title{Create an analyser from a searcher}
\usage{
Analyser$fromSearcher(searcher)
}
\arguments{
\item{searcher}{the name of the file - initial part of path - no .vocab.rds extension}

\item{expanded}{has the searcher already had ancestors expanded?}
}
\value{
an analyser with idf and graph counts precomputed.
}
\description{
creates a network analyser from a searcher
}
