% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{anyOf}
\alias{anyOf}
\title{convert a single row dataframe of values to a vector of values for use with %in%}
\usage{
anyOf(singleRowDf, ...)
}
\arguments{
\item{singleRowDf}{- a single rowed df of values}
}
\value{
vector of values
}
\description{
convert a single row dataframe of values to a vector of values for use with %in%
}
