% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{Cohort_fromDataframe}
\alias{Cohort_fromDataframe}
\title{Create a cohort from a data frame}
\usage{
Cohort$fromDataframe(df)
}
\arguments{
\item{omop}{the omop connection as a Omop class object}

\item{df}{a dataframe with a person_id field.}
}
\value{
a data from of targetConcepts
}
\description{
Create a cohort from a data frame
}
