% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{findStartDatetime}
\alias{findStartDatetime}
\title{check an omop related dataset for a start date}
\usage{
findStartDatetime(df)
}
\arguments{
\item{df}{- an omop derived df which may be a dbplyr table}
}
\value{
the symbolic representation of the first ??_start_datetime or ??_datetime column or NULL if there is none
}
\description{
check an omop related dataset for a start date
}
