% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searcher.R
\name{fromDataframe}
\alias{fromDataframe}
\title{Create a searcher from a data frame}
\usage{
Searcher$fromDataframe(omop,df,field)
}
\arguments{
\item{omop}{the omop connection as a Omop class object}

\item{df}{a dataframe with a ???_concept_id field}

\item{field}{the name of the field containing concept_ids}
}
\value{
a data from of concepts
}
\description{
selects concepts which are used in a specific vocabulary and returns the results
}
