% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualInformation.R
\name{mi}
\alias{mi}
\title{calculate mutual information from probabilities}
\usage{
mi(px1y1, px0y1, px1y0)
}
\arguments{
\item{px1y1}{probability of cooccurrence of x and y.}

\item{px0y1}{probability of occurrence of y without x}

\item{px1y0}{probability of occurrence of x without y}
}
\value{
the mutual information
}
\description{
calculate mutual information from probabilities
}
