% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualInformation.R
\name{npmi}
\alias{npmi}
\title{A calculate normalised pointwise mutual information from probabilities}
\usage{
npmi(pxy, px, py)
}
\arguments{
\item{pxy}{probability of cooccurrence of x and y.}

\item{px}{probability of occurrence of x}

\item{py}{probability of occurrence of y}
}
\value{
the normalised pointwise mutual information
}
\description{
A calculate normalised pointwise mutual information from probabilities
}
