% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omop.R
\name{Omop}
\alias{Omop}
\title{Get a handle to the omop databse}
\description{
Get a handle to the omop databse

Get a handle to the omop databse
}
\details{
This function sets up a connection to the omop databse.

N.B. if the df is a table to lookup is done in the database, otherwise the vocab is copied accross
}
\examples{

## ------------------------------------------------
## Method `Omop$new`
## ------------------------------------------------

omop = Omop$new(dbname,config)

## ------------------------------------------------
## Method `Omop$getConceptNames`
## ------------------------------------------------

omop = Omop$new()
df \%>\% omop$getConceptNames()
}
\keyword{omop}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{an connection the omop database}

\item{\code{resultsCon}}{an connection the omop results database}

\item{\code{resultsCon}}{an connection the omop results database}

\item{\code{care_site}}{omop table}

\item{\code{cdm_source}}{omop table}

\item{\code{concept}}{omop table}

\item{\code{concept_ancestor}}{omop table}

\item{\code{concept_class}}{omop table}

\item{\code{concept_relationship}}{omop table}

\item{\code{concept_synonym}}{omop table}

\item{\code{condition_era}}{omop table}

\item{\code{condition_occurrence}}{omop table}

\item{\code{cost}}{omop table}

\item{\code{device_exposure}}{omop table}

\item{\code{domain}}{omop table}

\item{\code{dose_era}}{omop table}

\item{\code{drug_era}}{omop table}

\item{\code{drug_exposure}}{omop table}

\item{\code{drug_strength}}{omop table}

\item{\code{fact_relationship}}{omop table}

\item{\code{location}}{omop table}

\item{\code{location_history}}{omop table}

\item{\code{measurement}}{omop table}

\item{\code{metadata}}{omop table}

\item{\code{note}}{omop table}

\item{\code{note_nlp}}{omop table}

\item{\code{observation}}{omop table}

\item{\code{observation_period}}{omop table}

\item{\code{payer_plan_period}}{omop table}

\item{\code{person}}{omop table}

\item{\code{procedure_occurrence}}{omop table}

\item{\code{provider}}{omop table}

\item{\code{relationship}}{omop table}

\item{\code{source_to_concept_map}}{omop table}

\item{\code{specimen}}{omop table}

\item{\code{survey_conduct}}{omop table}

\item{\code{visit_detail}}{omop table}

\item{\code{visit_occurrence}}{omop table}

\item{\code{vocabulary}}{omop table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Omop$new()}}
\item \href{#method-finalize}{\code{Omop$finalize()}}
\item \href{#method-forceCalculation}{\code{Omop$forceCalculation()}}
\item \href{#method-useCached}{\code{Omop$useCached()}}
\item \href{#method-persist}{\code{Omop$persist()}}
\item \href{#method-buildVocabSet}{\code{Omop$buildVocabSet()}}
\item \href{#method-buildCohort}{\code{Omop$buildCohort()}}
\item \href{#method-getConceptNames}{\code{Omop$getConceptNames()}}
\item \href{#method-expandAncestorConcepts}{\code{Omop$expandAncestorConcepts()}}
\item \href{#method-getSpanningGraphEdges}{\code{Omop$getSpanningGraphEdges()}}
\item \href{#method-getVisits}{\code{Omop$getVisits()}}
\item \href{#method-getDemographics}{\code{Omop$getDemographics()}}
\item \href{#method-getObserations}{\code{Omop$getObserations()}}
\item \href{#method-getProcedures}{\code{Omop$getProcedures()}}
\item \href{#method-getMeasurements}{\code{Omop$getMeasurements()}}
\item \href{#method-getConditions}{\code{Omop$getConditions()}}
\item \href{#method-getNotes}{\code{Omop$getNotes()}}
\item \href{#method-getNoteNlp}{\code{Omop$getNoteNlp()}}
\item \href{#method-getDrugs}{\code{Omop$getDrugs()}}
\item \href{#method-getDevices}{\code{Omop$getDevices()}}
\item \href{#method-getDeviceConcepts}{\code{Omop$getDeviceConcepts()}}
\item \href{#method-getDrugConcepts}{\code{Omop$getDrugConcepts()}}
\item \href{#method-getNoteNlpConcepts}{\code{Omop$getNoteNlpConcepts()}}
\item \href{#method-getConditionConcepts}{\code{Omop$getConditionConcepts()}}
\item \href{#method-getMeasurementConcepts}{\code{Omop$getMeasurementConcepts()}}
\item \href{#method-getMeasurementValueConcepts}{\code{Omop$getMeasurementValueConcepts()}}
\item \href{#method-getProcedureConcepts}{\code{Omop$getProcedureConcepts()}}
\item \href{#method-getProcedureModifierConcepts}{\code{Omop$getProcedureModifierConcepts()}}
\item \href{#method-getObservationConcepts}{\code{Omop$getObservationConcepts()}}
\item \href{#method-inferType}{\code{Omop$inferType()}}
\item \href{#method-clone}{\code{Omop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Sets up an omop database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$new(
  dbname = "omop",
  resultsdbname = "omopResults",
  config = config::get(file = "~/Dropbox/db.yaml")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbname}}{String value of dbname defaults to "omop".}

\item{\code{resultsdbname}}{String value of dbname defaults to "omopResults".}

\item{\code{config}}{Config object from yaml file. (defaults to config::get(file="~/Dropbox/db.yaml"))}

\item{\code{con}}{an ODBC connection to the omop database}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new(dbname,config)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
Closes the omop database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forceCalculation"></a>}}
\subsection{Method \code{forceCalculation()}}{
force recaluclation of cached results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$forceCalculation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-useCached"></a>}}
\subsection{Method \code{useCached()}}{
persist a dbplyr result in the omopResults database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$useCached()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persist"></a>}}
\subsection{Method \code{persist()}}{
persist a dbplyr result in the omopResults database.
Do not combine with compute(). MS SQL specific code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$persist(tableName, remoteDf, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to dbplyr::copy_to - of particular interest is unique_indexes = list("x","y"), indexes = list("x","y"), analyze = TRUE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildVocabSet"></a>}}
\subsection{Method \code{buildVocabSet()}}{
create a new vocabulary specification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$buildVocabSet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a searcher
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCohort"></a>}}
\subsection{Method \code{buildCohort()}}{
create a new cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$buildCohort(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a very short descriptive name for the cohort}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a cohort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConceptNames"></a>}}
\subsection{Method \code{getConceptNames()}}{
resolve all ???_concept_id fields in result set and look them up in the concept table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getConceptNames(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{an input data frame (or dbplyr lazy table)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new()
df \%>\% omop$getConceptNames()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expandAncestorConcepts"></a>}}
\subsection{Method \code{expandAncestorConcepts()}}{
get ancestors for a grouped set of concepts preserving grouping and count information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$expandAncestorConcepts(groupedDf, conceptIdVar, countVar = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupedDf}}{a grouped data frame}

\item{\code{conceptIdVar}}{the concept variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the analyser with expanded groupedDf including extra concepts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSpanningGraphEdges"></a>}}
\subsection{Method \code{getSpanningGraphEdges()}}{
for a set of concepts get all the hierarchical relationships between all concept pairs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getSpanningGraphEdges(df, conceptIdVar, min = 0L, max = 1000L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{a data frame containing concept ids}

\item{\code{conceptIdVar}}{the concept variable}

\item{\code{min}}{the minimum distance to consider}

\item{\code{max}}{the maximum distance to consider}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a set of edges with
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVisits"></a>}}
\subsection{Method \code{getVisits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getVisits(visit_type = omopCodes$visit_type \%>\% anyOf(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDemographics"></a>}}
\subsection{Method \code{getDemographics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getDemographics(gender = omopCodes$gender \%>\% anyOf(), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getObserations"></a>}}
\subsection{Method \code{getObserations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getObserations(conceptSet = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcedures"></a>}}
\subsection{Method \code{getProcedures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getProcedures(conceptSet = NULL, modfierConceptSet = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeasurements"></a>}}
\subsection{Method \code{getMeasurements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getMeasurements(conceptSet = NULL, valueConceptSet = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditions"></a>}}
\subsection{Method \code{getConditions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getConditions(conceptSet = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNotes"></a>}}
\subsection{Method \code{getNotes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getNotes(note_type = NULL, note_class = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNoteNlp"></a>}}
\subsection{Method \code{getNoteNlp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getNoteNlp(
  conceptSet = NULL,
  note_type = NULL,
  note_class = NULL,
  note_nlp_temporal = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDrugs"></a>}}
\subsection{Method \code{getDrugs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getDrugs(conceptSet = NULL, drug_route = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDevices"></a>}}
\subsection{Method \code{getDevices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getDevices(conceptSet = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDeviceConcepts"></a>}}
\subsection{Method \code{getDeviceConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getDeviceConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDrugConcepts"></a>}}
\subsection{Method \code{getDrugConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getDrugConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNoteNlpConcepts"></a>}}
\subsection{Method \code{getNoteNlpConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getNoteNlpConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionConcepts"></a>}}
\subsection{Method \code{getConditionConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getConditionConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeasurementConcepts"></a>}}
\subsection{Method \code{getMeasurementConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getMeasurementConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeasurementValueConcepts"></a>}}
\subsection{Method \code{getMeasurementValueConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getMeasurementValueConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcedureConcepts"></a>}}
\subsection{Method \code{getProcedureConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getProcedureConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcedureModifierConcepts"></a>}}
\subsection{Method \code{getProcedureModifierConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getProcedureModifierConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getObservationConcepts"></a>}}
\subsection{Method \code{getObservationConcepts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getObservationConcepts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inferType"></a>}}
\subsection{Method \code{inferType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$inferType(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
