% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{plot.Cohort}
\alias{plot.Cohort}
\title{plots a Cohort object}
\usage{
\method{plot}{Cohort}(x, smoothing = 20, binwidth = 5, ...)
}
\arguments{
\item{x}{- a cohort}

\item{smoothing}{- the half width of the smoothing window to apply to the date  in measurements}

\item{binwidth}{- the width in years of each age group in the histogram}
}
\description{
plots a Cohort object
}
