% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualInformation.R
\name{pmi}
\alias{pmi}
\title{A calculate pointwise mutual information from probabilities}
\usage{
pmi(pxy, px, py)
}
\arguments{
\item{pxy}{probability of cooccurrence of x and y.}

\item{px}{probability of occurrence of x}

\item{py}{probability of occurrence of y}
}
\value{
the pointwise mutual information
}
\description{
A calculate pointwise mutual information from probabilities
}
