% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searcher.R
\name{Searcher_fromConceptCode}
\alias{Searcher_fromConceptCode}
\title{Create a searcher from a code based search}
\usage{
Searcher$fromConceptCode(omop,vocabulary_id,concept_code)
}
\arguments{
\item{omop}{the omop connection as a Omop class object}

\item{vocabulary_id}{a vocab filter (can take sql wild card)}

\item{concept_code}{a concept_code (can take sql wild card)}
}
\value{
a searcher of matching concepts
}
\description{
get concept by vocabulary and concept code and return as a searcher
}
