% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searcher.R
\name{Searcher_fromSearch}
\alias{Searcher_fromSearch}
\title{Create a searcher from a search term}
\usage{
Searcher$fromSearch(omop,term)
}
\arguments{
\item{omop}{a object of class Omop}

\item{term}{an input search string}
}
\value{
a data.frame of domain_id and counts
}
\description{
perform a search of the concept table
}
