% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searcher.R
\name{Searcher_load}
\alias{Searcher_load}
\title{Create a searcher from a .RDS file}
\usage{
Searcher$load(path)
}
\arguments{
\item{name}{the name of the file - initial part of path - no .vocab.rds extension}
}
\value{
the modified searcher itself
}
\description{
loads the data from local path
}
