% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{typedTbl}
\alias{typedTbl}
\title{#' Normalise an omop dataframe to a consistent format to construct a single feature set accross
#' 
#' with the following columns: 
#' 
#' * cohort_person_id, 
#' * cohort_entry_datetime, 
#' * feature_source, e.g. measurement, observation, etc...
#' * feature_name, 
#' * feature_concept_id, 
#' * feature_value_as_number, 
#' * feature_value_as_date, 
#' * feature_value_as_name, 
#' * feature_value_as_concept_id, 
#' * feature_days_offset,
#' * feature_display
#' 
#' or the equivalent with observation_ prefix
#' 
#' @param omopDf - a df which may be a dbplyr table
#' @param outcome - is the 
#' @return a dbplyr dataframe 
normaliseColumns = function(omopDf, prefix = "feature") {
  omopDf = omopDf %>% omop$inferType()
  if (""class(omopDf)
  omopDf = omopDf %>% mutate(prefix_source = ) %>%
    rename(
      prefix_datetime = any_of(c("")),
      prefix_concept_id = any_of(c("")),
      prefix_value_as_number = any_of(c("")),
      prefix_value_as_number_min = any_of(c("")),
      prefix_value_as_number_max = any_of(c("")),
      prefix_value_as_number_unit = any_of(c("")),
      prefix_value_as_date = any_of(c("")),
      prefix_value_as_name = any_of(c("")),
      prefix_value_as_concept_id = any_of(c("")),
      prefix_days_offset = days_offset,
      prefix_display = c()
    )
  replaceList = colnames(omopDf) %>% stringr::str_replace("prefix",prefix)
  replaceList = replaceList[replaceList != colnames(omopDf)]
  names(replaceList) = colnames(omopDf)
  omopDf = omopDf %>% rename(replaceList)
  return(omopDf)
  #x = tibble(y=c(1,2,3),w=c(3,2,1))
  #x %>% select(y,z=any_of(c("a","c")))#,"w","d","y")))
}}
\usage{
typedTbl(con, tableName)
}
\arguments{
\item{con}{- an omop database connection}

\item{tableName}{- the table name}
}
\value{
the dbplyr table with addition class information
}
\description{
c("care_site","cdm_source","concept","concept_ancestor","concept_class","concept_relationship","concept_synonym","condition_era",
  "condition_occurrence","cost","device_exposure","domain","dose_era","drug_era","drug_exposure","drug_strength",
  "fact_relationship","location","location_history","measurement","metadata","note","note_nlp","observation","observation_period",
  "payer_plan_period","persist","persistance","person","procedure_occurrence","provider","relationship","source_to_concept_map",
  "specimen","survey_conduct","visit_detail","visit_occurrence","vocabulary")
}
\details{
a dbplyr table with addition class information
}
