% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{Analyser}
\alias{Analyser}
\title{Get an of the omop databse analyser}
\description{
Get an of the omop databse analyser

Get an of the omop databse analyser
}
\examples{

## ------------------------------------------------
## Method `Analyser$calculateTfidf`
## ------------------------------------------------

omop = Omop$new(...)
a = Analyser$new(omop)
df = ... get all concepts from a set of documents ...

## ------------------------------------------------
## Method `Analyser$plotConceptNetwork`
## ------------------------------------------------

omop = Omop$new(...)
s = Searcher$new(omop)
df = s.search("\%myoc\%inf\%") 
a = Analyser$from...
a$plotConceptNetwork()
}
\keyword{omop}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{omop}}{the omop connection}

\item{\code{groupedDf}}{a grouped dataframe including a grouping representing the documents, distinct concepts in the groups and counts of concepts in document}

\item{\code{grps}}{the original grouping of the data - corresponds to a document}

\item{\code{nodes}}{a dataframe including unique concept_ids, overall counts and count of documents containing}

\item{\code{edges}}{a dataframe of shortest ancestor edges between nodes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analyser$new()}}
\item \href{#method-countAncestorConcepts}{\code{Analyser$countAncestorConcepts()}}
\item \href{#method-calculateTfidf}{\code{Analyser$calculateTfidf()}}
\item \href{#method-plotConceptNetwork}{\code{Analyser$plotConceptNetwork()}}
\item \href{#method-plotSankeyD3}{\code{Analyser$plotSankeyD3()}}
\item \href{#method-print}{\code{Analyser$print()}}
\item \href{#method-clone}{\code{Analyser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
create a concept analyser - typically not designed to be called directly. See Analyser$fromSearcher and Analyser$fromDataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$new(omop, field = "concept_id", expanded = FALSE, groupedDf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omop}}{an R6 Omop object}

\item{\code{field}}{the field name of the concept_id}

\item{\code{expanded}}{has the data already been expanded to include ancestors}

\item{\code{groupedDf}}{a data frame including at least a concept_id plus the groupings defining documents, and ideally counts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-countAncestorConcepts"></a>}}
\subsection{Method \code{countAncestorConcepts()}}{
get ancestors for a grouped set of concepts preserving grouping and count information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$countAncestorConcepts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the analyser with expanded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateTfidf"></a>}}
\subsection{Method \code{calculateTfidf()}}{
get an inverse document frequency for concepts associated with a grouping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateTfidf()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The grouped datafram here acts as a "Document" from the perpective of the TFIDF calculation but might be a person

a dataframe with a concept_id field and at least one group_by active defining the "document"
}

\subsection{Returns}{
an analysis object with the groupedDf expanded to include tfidf stats for each concept in each group
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new(...)
a = Analyser$new(omop)
df = ... get all concepts from a set of documents ...
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotConceptNetwork"></a>}}
\subsection{Method \code{plotConceptNetwork()}}{
plot the currently loaded graph.
TODO: facet plot on different scoreVars
TODO: make layout fixed for multiple graphs: https://github.com/thomasp85/ggraph/issues/130
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$plotConceptNetwork(scoreVar = "count", labelScore = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scoreVar}}{a node feature the the graph will use as e.g.count, norm_okapi_bm25,}

\item{\code{labelScore}}{a percentage of score that will trigger a node being labelled}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the networkAnalysis object with a graph loaded
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new(...)
s = Searcher$new(omop)
df = s.search("\%myoc\%inf\%") 
a = Analyser$from...
a$plotConceptNetwork()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSankeyD3"></a>}}
\subsection{Method \code{plotSankeyD3()}}{
plot the currently loaded graph as sankey diagram.
TODO: facet plot on different scoreVars
TODO: make layout fixed for multiple graphs: https://github.com/thomasp85/ggraph/issues/130
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$plotSankeyD3(linkWeightVar = "weight")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linkWeightVar}}{the score on the link that will define the width of the network}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print the analyser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
