% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysis.R
\name{Analyser}
\alias{Analyser}
\title{Get an of the omop databse analyser}
\description{
Get an of the omop databse analyser

Get an of the omop databse analyser
}
\examples{

## ------------------------------------------------
## Method `Analyser$plotConceptNetwork`
## ------------------------------------------------

omop = Omop$new(...)
s = Searcher$new(omop)
df = s.search("\%myoc\%inf\%") 
a = Analyser$from...
a$plotConceptNetwork()
}
\keyword{omop}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{omop}}{the omop connection}

\item{\code{groupedDf}}{a grouped dataframe including a grouping representing the documents, distinct concepts in the groups and counts of concepts in document}

\item{\code{grps}}{the original grouping of the data - corresponds to a document}

\item{\code{nodes}}{a dataframe including unique concept_ids, overall counts and count of documents containing}

\item{\code{edges}}{a dataframe of shortest ancestor edges between nodes}

\item{\code{totalDocuments}}{a count of all the docuemnt level groupings in the analyser}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analyser$new()}}
\item \href{#method-calculateIdfNodes}{\code{Analyser$calculateIdfNodes()}}
\item \href{#method-calculateGraph}{\code{Analyser$calculateGraph()}}
\item \href{#method-redefineGroups}{\code{Analyser$redefineGroups()}}
\item \href{#method-calculateAncestorConcepts}{\code{Analyser$calculateAncestorConcepts()}}
\item \href{#method-calculateOutcomeMI}{\code{Analyser$calculateOutcomeMI()}}
\item \href{#method-calculateTfidf}{\code{Analyser$calculateTfidf()}}
\item \href{#method-plotConceptNetwork}{\code{Analyser$plotConceptNetwork()}}
\item \href{#method-plotSankeyD3}{\code{Analyser$plotSankeyD3()}}
\item \href{#method-calculateCooccurrenceMI}{\code{Analyser$calculateCooccurrenceMI()}}
\item \href{#method-print}{\code{Analyser$print()}}
\item \href{#method-clone}{\code{Analyser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
create a concept analyser - typically not designed to be called directly. See Analyser$fromSearcher and Analyser$fromDataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$new(omop, groupedDf, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omop}}{an R6 Omop object}

\item{\code{groupedDf}}{a data frame including at least a concept_id plus the groupings defining documents, and ideally counts}

\item{\code{field}}{the field name of the concept_id}

\item{\code{expanded}}{has the data already been expanded to include ancestors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateIdfNodes"></a>}}
\subsection{Method \code{calculateIdfNodes()}}{
creates the nodes field containing IDF information

TODO: save and restore IDFs based on different corpuses
TODO: calculate IDFs of ancestor frequencies based on child frequencies (if possible)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateIdfNodes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateGraph"></a>}}
\subsection{Method \code{calculateGraph()}}{
build the nodes and links using the ancestor table and shortest paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateGraph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-redefineGroups"></a>}}
\subsection{Method \code{redefineGroups()}}{
define document aggregation grouping post hoc. usually better to group data before importing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$redefineGroups(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the analyser
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateAncestorConcepts"></a>}}
\subsection{Method \code{calculateAncestorConcepts()}}{
get ancestors for a grouped set of concepts preserving grouping and count information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateAncestorConcepts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the analyser with expanded groupedDf including extra concepts
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateOutcomeMI"></a>}}
\subsection{Method \code{calculateOutcomeMI()}}{
caluculates a mutual information table for the outcomes given in outcomeVar
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateOutcomeMI(outcomeVar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outcomeVar}}{the outcome}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The grouped dataframe here acts as a "Document" from the perpective of the TFIDF calculation but might be a person
}

\subsection{Returns}{
an analysis object with the nodes expanded to include a MI score for each concept assuming a binary outcome
and an expanded groupedDf which includes the normalised pointwise mutual information stats for each concept in each outcome
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateTfidf"></a>}}
\subsection{Method \code{calculateTfidf()}}{
get an inverse document frequency for concepts associated with a grouping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateTfidf(k1 = 1.2, b = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k1}}{default 1.2 - okapi BM25}

\item{\code{b}}{default 0.95 - okapi BM25}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The grouped datafram here acts as a "Document" from the perpective of the TFIDF calculation but might be a person
}

\subsection{Returns}{
an analysis object with the groupedDf expanded to include tfidf stats for each concept in each group
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotConceptNetwork"></a>}}
\subsection{Method \code{plotConceptNetwork()}}{
plot the currently loaded graph.
TODO: facet plot on different scoreVars
TODO: make layout fixed for multiple graphs: https://github.com/thomasp85/ggraph/issues/130
TODO: split this into own class?
TODO: aggregatre tfidf measures
TODO: mutual information measures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$plotConceptNetwork(scoreVar = "f_t", labelScore = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scoreVar}}{a node feature the the graph will use as e.g. f_t, N_t, idf measures...}

\item{\code{labelScore}}{a percentage of score that will trigger a node being labelled}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the networkAnalysis object with a graph loaded
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new(...)
s = Searcher$new(omop)
df = s.search("\%myoc\%inf\%") 
a = Analyser$from...
a$plotConceptNetwork()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSankeyD3"></a>}}
\subsection{Method \code{plotSankeyD3()}}{
plot the currently loaded graph as sankey diagram.
TODO: figure out how the links get a weight that is meaningful in this context
This would be good for visualising a classifier network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$plotSankeyD3(linkWeightVar = "weight")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linkWeightVar}}{the score on the link that will define the width of the network}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateCooccurrenceMI"></a>}}
\subsection{Method \code{calculateCooccurrenceMI()}}{
calculate a co-occurence matrix for the most
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$calculateCooccurrenceMI(filteredNodes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filteredNodes}}{- a list of}

\item{\code{limit}}{the max number of terms co-occurrences that will be used for cooccurrence or NA for no limit}

\item{\code{ascending}}{order the factor as ascending}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the co-occurrence matrix as a dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print the analyser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
