% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omop.R
\name{Omop}
\alias{Omop}
\title{Get a handle to the omop databse}
\description{
Get a handle to the omop databse

Get a handle to the omop databse
}
\details{
This function sets up a connection to the omop databse.

N.B. if the df is a table to lookup is done in the database, otherwise the vocab is copied accross
}
\examples{

## ------------------------------------------------
## Method `Omop$new`
## ------------------------------------------------

omop = Omop$new(dbname,config)

## ------------------------------------------------
## Method `Omop$getConceptNames`
## ------------------------------------------------

omop = Omop$new()
df \%>\% omop$getConceptNames()
}
\keyword{omop}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{an connection the omop database}

\item{\code{care_site}}{omop table}

\item{\code{cdm_source}}{omop table}

\item{\code{concept}}{omop table}

\item{\code{concept_ancestor}}{omop table}

\item{\code{concept_class}}{omop table}

\item{\code{concept_relationship}}{omop table}

\item{\code{concept_synonym}}{omop table}

\item{\code{condition_era}}{omop table}

\item{\code{condition_occurrence}}{omop table}

\item{\code{cost}}{omop table}

\item{\code{device_exposure}}{omop table}

\item{\code{domain}}{omop table}

\item{\code{dose_era}}{omop table}

\item{\code{drug_era}}{omop table}

\item{\code{drug_exposure}}{omop table}

\item{\code{drug_strength}}{omop table}

\item{\code{fact_relationship}}{omop table}

\item{\code{location}}{omop table}

\item{\code{location_history}}{omop table}

\item{\code{measurement}}{omop table}

\item{\code{metadata}}{omop table}

\item{\code{note}}{omop table}

\item{\code{note_nlp}}{omop table}

\item{\code{observation}}{omop table}

\item{\code{observation_period}}{omop table}

\item{\code{payer_plan_period}}{omop table}

\item{\code{person}}{omop table}

\item{\code{procedure_occurrence}}{omop table}

\item{\code{provider}}{omop table}

\item{\code{relationship}}{omop table}

\item{\code{source_to_concept_map}}{omop table}

\item{\code{specimen}}{omop table}

\item{\code{survey_conduct}}{omop table}

\item{\code{visit_detail}}{omop table}

\item{\code{visit_occurrence}}{omop table}

\item{\code{vocabulary}}{omop table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Omop$new()}}
\item \href{#method-finalize}{\code{Omop$finalize()}}
\item \href{#method-buildVocabSet}{\code{Omop$buildVocabSet()}}
\item \href{#method-buildCohort}{\code{Omop$buildCohort()}}
\item \href{#method-getConceptNames}{\code{Omop$getConceptNames()}}
\item \href{#method-clone}{\code{Omop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Sets up an omop database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$new(dbname = "omop", config = config::get(file = "~/Dropbox/db.yaml"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbname}}{String value of dbname defaults to "omop".}

\item{\code{config}}{Config object from yaml file. (defaults to config::get(file="~/Dropbox/db.yaml"))}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new(dbname,config)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
Closes the omop database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildVocabSet"></a>}}
\subsection{Method \code{buildVocabSet()}}{
create a new vocabulary specification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$buildVocabSet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a searcher
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildCohort"></a>}}
\subsection{Method \code{buildCohort()}}{
create a new cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$buildCohort(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a very short descriptive name for the cohort}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a cohort
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConceptNames"></a>}}
\subsection{Method \code{getConceptNames()}}{
resolve all ???_concept_id fields in result set and look them up in the concept table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$getConceptNames(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{an input data frame (or dbplyr lazy table)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{omop = Omop$new()
df \%>\% omop$getConceptNames()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Omop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
